Feat: detailPageButton
