+++ Sojumap/View/MapViewController.storyboard
@@ -4,6 +4,7 @@
     <dependencies>
         <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21679"/>
+        <capability name="Image references" minToolsVersion="12.0"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="System colors in document resources" minToolsVersion="11.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
@@ -18,20 +19,37 @@
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="dzo-KA-DnH" customClass="NMFNaverMapView">
-                                <rect key="frame" x="0.0" y="0.0" width="393" height="669"/>
+                                <rect key="frame" x="0.0" y="0.0" width="393" height="639"/>
                                 <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                             </view>
                             <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="8" translatesAutoresizingMaskIntoConstraints="NO" id="gVo-6y-TiB">
-                                <rect key="frame" x="16" y="685" width="360" height="50.333333333333371"/>
+                                <rect key="frame" x="16" y="659" width="360" height="61.666666666666629"/>
                                 <subviews>
-                                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="식당 이름" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="XLP-2h-BCe" userLabel="placeName">
-                                        <rect key="frame" x="0.0" y="0.0" width="360" height="23"/>
-                                        <fontDescription key="fontDescription" type="system" weight="semibold" pointSize="19"/>
-                                        <nil key="textColor"/>
-                                        <nil key="highlightedColor"/>
-                                    </label>
+                                    <stackView opaque="NO" contentMode="scaleToFill" spacing="8" translatesAutoresizingMaskIntoConstraints="NO" id="xeq-lw-MFa">
+                                        <rect key="frame" x="0.0" y="0.0" width="360" height="34.333333333333336"/>
+                                        <subviews>
+                                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="식당 이름" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="XLP-2h-BCe" userLabel="placeName">
+                                                <rect key="frame" x="0.0" y="0.0" width="302" height="34.333333333333336"/>
+                                                <fontDescription key="fontDescription" type="system" weight="semibold" pointSize="22"/>
+                                                <nil key="textColor"/>
+                                                <nil key="highlightedColor"/>
+                                            </label>
+                                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="V9K-EU-65X">
+                                                <rect key="frame" x="310" y="0.0" width="50" height="34.333333333333336"/>
+                                                <constraints>
+                                                    <constraint firstAttribute="width" constant="50" id="82H-XD-2WL"/>
+                                                </constraints>
+                                                <buttonConfiguration key="configuration" style="plain">
+                                                    <imageReference key="image" image="info.circle" catalog="system" symbolScale="large"/>
+                                                </buttonConfiguration>
+                                                <connections>
+                                                    <action selector="detailPageButton:" destination="Y6W-OH-hqX" eventType="touchUpInside" id="7Hf-K5-ghz"/>
+                                                </connections>
+                                            </button>
+                                        </subviews>
+                                    </stackView>
                                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="주소" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="l7B-MH-Bse" userLabel="placeAddr">
-                                        <rect key="frame" x="0.0" y="31" width="360" height="19.333333333333329"/>
+                                        <rect key="frame" x="0.0" y="42.333333333333371" width="360" height="19.333333333333329"/>
                                         <constraints>
                                             <constraint firstAttribute="width" constant="360" id="GEn-Fy-HXt"/>
                                         </constraints>
@@ -46,8 +64,8 @@
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
                             <constraint firstItem="gVo-6y-TiB" firstAttribute="leading" secondItem="vDu-zF-Fre" secondAttribute="leading" constant="16" id="6MJ-sE-t9H"/>
-                            <constraint firstItem="gVo-6y-TiB" firstAttribute="top" secondItem="dzo-KA-DnH" secondAttribute="bottom" constant="16" id="JTr-00-ijX"/>
-                            <constraint firstItem="vDu-zF-Fre" firstAttribute="bottom" secondItem="dzo-KA-DnH" secondAttribute="bottom" constant="100" id="KTg-3u-NBe"/>
+                            <constraint firstItem="gVo-6y-TiB" firstAttribute="top" secondItem="dzo-KA-DnH" secondAttribute="bottom" constant="20" id="JTr-00-ijX"/>
+                            <constraint firstItem="vDu-zF-Fre" firstAttribute="bottom" secondItem="dzo-KA-DnH" secondAttribute="bottom" constant="130" id="KTg-3u-NBe"/>
                             <constraint firstItem="dzo-KA-DnH" firstAttribute="trailing" secondItem="vDu-zF-Fre" secondAttribute="trailing" id="c1w-ub-Yuj"/>
                             <constraint firstItem="dzo-KA-DnH" firstAttribute="top" secondItem="5EZ-qb-Rvc" secondAttribute="top" id="oGt-g1-D7I"/>
                             <constraint firstItem="dzo-KA-DnH" firstAttribute="leading" secondItem="vDu-zF-Fre" secondAttribute="leading" id="p4k-5f-Nnd"/>
@@ -68,6 +86,7 @@
         </scene>
     </scenes>
     <resources>
+        <image name="info.circle" catalog="system" width="128" height="123"/>
         <image name="map" catalog="system" width="128" height="112"/>
         <systemColor name="systemBackgroundColor">
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
